updating schema for dirty-db and writing to tempfile before sending to s3

+++ serializer/dirty-db-api/src/main/sql/schema.sql
@@ -1,11 +1,11 @@
 drop table dirty_list;
 
 create table dirty_list(
-  sno int auto_increment primary key,
+  sno bigint auto_increment primary key,
   id varchar(20),
   `type` varchar(20),
   country varchar(2),
   language varchar(2),
   lastmodified timestamp,
-  sequencenumber int
+  sequencenumber bigint
 );
